Stops the scripts explicitly.
When there are no files to process on, we should stop the script
directly. So other parts of the script can expect that there are files
to operate on. With the last commit, this is needed, as we execute
next(iterator), that fails if there is no element in the iterator.