Revert "Fix race condition of $XDG_CACHE_HOME/rply creation." (refs: #17359).
This reverts commit c957b5b8814778bdcb315d181e23095d1d7cb6bc.

With the new approach (implemented two commits ago), a cache directory
is created for each i18nspector call. This means that we don't need to
treat the first file specifically (which wasn't working, depending on
which file that was).

As a result, the code can be simplified, getting back to the previous
“treat all files the same way” implementation.